requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
Flask==3.1.0
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.1
pymongo==4.13.0
python-dotenv==1.1.0
rich==13.9.4
typing_extensions==4.14.0
Werkzeug==3.1.3
wrapt==1.17.2
